feat(detection): several enhancement and reformat
