+++ hubconf.py
@@ -1,3 +1,14 @@
+from official.nlp.bert.model import (
+    cased_L_12_H_768_A_12,
+    cased_L_24_H_1024_A_16,
+    chinese_L_12_H_768_A_12,
+    multi_cased_L_12_H_768_A_12,
+    uncased_L_12_H_768_A_12,
+    uncased_L_24_H_1024_A_16,
+    wwm_cased_L_24_H_1024_A_16,
+    wwm_uncased_L_24_H_1024_A_16,
+)
+from official.quantization.models import quantized_resnet18
 from official.vision.classification.resnet.model import (
     BasicBlock,
     Bottleneck,
@@ -16,54 +27,22 @@ from official.vision.classification.shufflenet.model import (
     shufflenet_v2_x1_5,
     shufflenet_v2_x2_0,
 )
-
-from official.nlp.bert.model import (
-    uncased_L_12_H_768_A_12,
-    cased_L_12_H_768_A_12,
-    uncased_L_24_H_1024_A_16,
-    cased_L_24_H_1024_A_16,
-    chinese_L_12_H_768_A_12,
-    multi_cased_L_12_H_768_A_12,
-    wwm_uncased_L_24_H_1024_A_16,
-    wwm_cased_L_24_H_1024_A_16,
-)
-
-from official.vision.detection.faster_rcnn_fpn_res50_coco_1x_800size import (
-    faster_rcnn_fpn_res50_coco_1x_800size,
-)
-from official.vision.detection.faster_rcnn_fpn_res50_coco_1x_800size_syncbn import (
-    faster_rcnn_fpn_res50_coco_1x_800size_syncbn,
-)
-
-from official.vision.detection.retinanet_res50_coco_1x_800size import (
+from official.vision.detection.configs import (
+    faster_rcnn_res50_coco_1x_800size,
+    faster_rcnn_res50_coco_1x_800size_syncbn,
     retinanet_res50_coco_1x_800size,
-)
-from official.vision.detection.retinanet_res50_coco_1x_800size_syncbn import (
     retinanet_res50_coco_1x_800size_syncbn,
 )
-# TODO: need pretrained weights
-# from official.vision.detection.retinanet_res50_objects365_1x_800size import (
-#     retinanet_res50_objects365_1x_800size,
-# )
-# from official.vision.detection.retinanet_res50_voc_1x_800size import (
-#     retinanet_res50_voc_1x_800size,
-# )
-
 from official.vision.detection.models import FasterRCNN, RetinaNet
-from official.vision.detection.tools.test import DetEvaluator
-
-from official.vision.segmentation.deeplabv3plus import (
-    deeplabv3plus_res101,
-    DeepLabV3Plus,
-)
-
+from official.vision.detection.tools.utils import DetEvaluator
+from official.vision.keypoints.inference import KeypointEvaluator
 from official.vision.keypoints.models import (
+    mspn_4stage,
     simplebaseline_res50,
     simplebaseline_res101,
     simplebaseline_res152,
-    mspn_4stage
 )
-
-from official.vision.keypoints.inference import KeypointEvaluator
-
-from official.quantization.models import quantized_resnet18
+from official.vision.segmentation.deeplabv3plus import (
+    DeepLabV3Plus,
+    deeplabv3plus_res101,
+)
