+++ official/vision/detection/configs/faster_rcnn_res50_coco_1x_800size_syncbn.py
@@ -11,7 +11,7 @@ from megengine import hub
 from official.vision.detection import models
 
 
-class CustomFasterRCNNFPNConfig(models.FasterRCNNConfig):
+class CustomFasterRCNNConfig(models.FasterRCNNConfig):
     def __init__(self):
         super().__init__()
 
@@ -24,7 +24,7 @@ class CustomFasterRCNNFPNConfig(models.FasterRCNNConfig):
     "https://data.megengine.org.cn/models/weights/"
     "faster_rcnn_fpn_cf5c020b_res50_1x_800size_syncbn_37dot6.pkl"
 )
-def faster_rcnn_fpn_res50_coco_1x_800size_syncbn(batch_size=1, **kwargs):
+def faster_rcnn_res50_coco_1x_800size_syncbn(batch_size=1, **kwargs):
     r"""
     Faster-RCNN FPN trained from COCO dataset.
     `"Faster-RCNN" <https://arxiv.org/abs/1506.01497>`_
@@ -32,8 +32,8 @@ def faster_rcnn_fpn_res50_coco_1x_800size_syncbn(batch_size=1, **kwargs):
     `"COCO" <https://arxiv.org/abs/1405.0312>`_
     `"SyncBN" <https://arxiv.org/abs/1711.07240>`_
     """
-    return models.FasterRCNN(CustomFasterRCNNFPNConfig(), batch_size=batch_size, **kwargs)
+    return models.FasterRCNN(CustomFasterRCNNConfig(), batch_size=batch_size, **kwargs)
 
 
 Net = models.FasterRCNN
-Cfg = CustomFasterRCNNFPNConfig
+Cfg = CustomFasterRCNNConfig
