+++ official/vision/detection/configs/retinanet_res50_coco_1x_800size.py
