+++ official/vision/detection/configs/retinanet_res50_objects365_1x_800size.py
@@ -6,8 +6,6 @@
 # Unless required by applicable law or agreed to in writing,
 # software distributed under the License is distributed on an
 # "AS IS" BASIS, WITHOUT ARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-from megengine import hub
-
 from official.vision.detection import models
 
 
@@ -17,25 +15,28 @@ class CustomRetinaNetConfig(models.RetinaNetConfig):
 
         # ------------------------ data cfg -------------------------- #
         self.train_dataset = dict(
-            name="voc",
-            root="VOCdevkit/VOC2012",
-            image_set="train",
+            name="objects365",
+            root="train",
+            ann_file="annotations/objects365_train_20190423.json",
+            remove_images_without_annotations=True,
         )
         self.test_dataset = dict(
-            name="voc",
-            root="VOCdevkit/VOC2012",
-            image_set="val",
+            name="objects365",
+            root="val",
+            ann_file="annotations/objects365_val_20190423.json",
+            remove_images_without_annotations=False,
         )
-        self.num_classes = 20
+        self.num_classes = 365
 
         # ------------------------ training cfg ---------------------- #
-        self.nr_images_epoch = 16000
+        self.nr_images_epoch = 400000
 
 
-def retinanet_res50_voc_1x_800size(batch_size=1, **kwargs):
+def retinanet_res50_objects365_1x_800size(batch_size=1, **kwargs):
     r"""
-    RetinaNet trained from VOC dataset.
+    RetinaNet trained from Objects365 dataset.
     `"RetinaNet" <https://arxiv.org/abs/1708.02002>`_
+    `"FPN" <https://arxiv.org/abs/1612.03144>`_
     """
     return models.RetinaNet(CustomRetinaNetConfig(), batch_size=batch_size, **kwargs)
