+++ official/vision/detection/layers/det/anchor.py
@@ -111,7 +111,7 @@ class DefaultAnchorGenerator(BaseAnchorGenerator):
         flatten_shift_y = F.add_axis(broad_shift_y.reshape(-1), 1)
 
         centers = F.concat(
-            [flatten_shift_x, flatten_shift_y, flatten_shift_x, flatten_shift_y, ],
+            [flatten_shift_x, flatten_shift_y, flatten_shift_x, flatten_shift_y,],
             axis=1,
         )
         centers = centers + self.offset * self.base_size
