+++ official/vision/detection/layers/det/rcnn.py
@@ -14,14 +14,10 @@ from official.vision.detection import layers
 
 
 class RCNN(M.Module):
-
     def __init__(self, cfg):
         super().__init__()
         self.cfg = cfg
-        self.box_coder = layers.BoxCoder(
-            reg_mean=cfg.rcnn_reg_mean,
-            reg_std=cfg.rcnn_reg_std
-        )
+        self.box_coder = layers.BoxCoder(cfg.rcnn_reg_mean, cfg.rcnn_reg_std)
 
         # roi head
         self.in_features = cfg.rcnn_in_features
@@ -44,12 +40,13 @@ class RCNN(M.Module):
             M.init.fill_(l.bias, 0)
 
     def forward(self, fpn_fms, rcnn_rois, im_info=None, gt_boxes=None):
-        rcnn_rois, labels, bbox_targets = self.get_ground_truth(rcnn_rois, im_info, gt_boxes)
+        rcnn_rois, labels, bbox_targets = self.get_ground_truth(
+            rcnn_rois, im_info, gt_boxes
+        )
 
         fpn_fms = [fpn_fms[x] for x in self.in_features]
         pool_features = layers.roi_pool(
-            fpn_fms, rcnn_rois, self.stride,
-            self.pooling_size, self.pooling_method,
+            fpn_fms, rcnn_rois, self.stride, self.pooling_size, self.pooling_method,
         )
         flatten_feature = F.flatten(pool_features, start_axis=1)
         roi_feature = F.relu(self.fc1(flatten_feature))
@@ -67,14 +64,13 @@ class RCNN(M.Module):
             pred_delta = F.indexing_one_hot(pred_delta, vlabels, axis=1)
 
             loss_rcnn_loc = layers.get_smooth_l1_loss(
-                pred_delta, bbox_targets, labels,
+                pred_delta,
+                bbox_targets,
+                labels,
                 self.cfg.rcnn_smooth_l1_beta,
                 norm_type="all",
             )
-            loss_dict = {
-                'loss_rcnn_cls': loss_rcnn_cls,
-                'loss_rcnn_loc': loss_rcnn_loc
-            }
+            loss_dict = {"loss_rcnn_cls": loss_rcnn_cls, "loss_rcnn_loc": loss_rcnn_loc}
             return loss_dict
         else:
             # slice 1 for removing background
@@ -82,7 +78,9 @@ class RCNN(M.Module):
             pred_delta = pred_delta[:, 4:].reshape(-1, 4)
             target_shape = (rcnn_rois.shapeof(0), self.cfg.num_classes, 4)
             # rois (N, 4) -> (N, 1, 4) -> (N, 80, 4) -> (N * 80, 4)
-            base_rois = F.add_axis(rcnn_rois[:, 1:5], 1).broadcast(target_shape).reshape(-1, 4)
+            base_rois = (
+                F.add_axis(rcnn_rois[:, 1:5], 1).broadcast(target_shape).reshape(-1, 4)
+            )
             pred_bbox = self.box_coder.decode(base_rois, pred_delta)
             return pred_bbox, pred_scores
 
@@ -101,7 +99,7 @@ class RCNN(M.Module):
             batch_inds = mge.ones((gt_boxes_per_img.shapeof(0), 1)) * bid
             # if config.proposal_append_gt:
             gt_rois = F.concat([batch_inds, gt_boxes_per_img[:, :4]], axis=1)
-            batch_roi_mask = (rpn_rois[:, 0] == bid)
+            batch_roi_mask = rpn_rois[:, 0] == bid
             _, batch_roi_inds = F.cond_take(batch_roi_mask == 1, batch_roi_mask)
             # all_rois : [batch_id, x1, y1, x2, y2]
             all_rois = F.concat([rpn_rois.ai[batch_roi_inds], gt_rois])
@@ -117,22 +115,26 @@ class RCNN(M.Module):
             gt_assignment_ignore = F.argmax(overlaps_ignore, axis=1)
 
             ignore_assign_mask = (max_overlaps_normal < self.cfg.fg_threshold) * (
-                max_overlaps_ignore > max_overlaps_normal)
+                max_overlaps_ignore > max_overlaps_normal
+            )
             max_overlaps = (
-                max_overlaps_normal * (1 - ignore_assign_mask) +
-                max_overlaps_ignore * ignore_assign_mask
+                max_overlaps_normal * (1 - ignore_assign_mask)
+                + max_overlaps_ignore * ignore_assign_mask
             )
             gt_assignment = (
-                gt_assignment_normal * (1 - ignore_assign_mask) +
-                gt_assignment_ignore * ignore_assign_mask
+                gt_assignment_normal * (1 - ignore_assign_mask)
+                + gt_assignment_ignore * ignore_assign_mask
             )
             gt_assignment = gt_assignment.astype("int32")
             labels = gt_boxes_per_img.ai[gt_assignment, 4]
 
             # ---------------- get the fg/bg labels for each roi ---------------#
-            fg_mask = (max_overlaps >= self.cfg.fg_threshold) * (labels != self.cfg.ignore_label)
+            fg_mask = (max_overlaps >= self.cfg.fg_threshold) * (
+                labels != self.cfg.ignore_label
+            )
             bg_mask = (max_overlaps < self.cfg.bg_threshold_high) * (
-                max_overlaps >= self.cfg.bg_threshold_low)
+                max_overlaps >= self.cfg.bg_threshold_low
+            )
 
             num_fg_rois = self.cfg.num_rois * self.cfg.fg_ratio
 
@@ -145,7 +147,7 @@ class RCNN(M.Module):
             keep_mask = fg_inds_mask + bg_inds_mask
             _, keep_inds = F.cond_take(keep_mask == 1, keep_mask)
             # Add next line to avoid memory exceed
-            keep_inds = keep_inds[:F.minimum(self.cfg.num_rois, keep_inds.shapeof(0))]
+            keep_inds = keep_inds[: F.minimum(self.cfg.num_rois, keep_inds.shapeof(0))]
             # labels
             labels = labels.ai[keep_inds].astype("int32")
             rois = all_rois.ai[keep_inds]
@@ -160,12 +162,12 @@ class RCNN(M.Module):
         return (
             F.zero_grad(F.concat(return_rois, axis=0)),
             F.zero_grad(F.concat(return_labels, axis=0)),
-            F.zero_grad(F.concat(return_bbox_targets, axis=0))
+            F.zero_grad(F.concat(return_bbox_targets, axis=0)),
         )
 
     def _bernoulli_sample_masks(self, masks, num_samples, sample_value):
         """ Using the bernoulli sampling method"""
-        sample_mask = (masks == sample_value)
+        sample_mask = masks == sample_value
         num_mask = sample_mask.sum()
         num_final_samples = F.minimum(num_mask, num_samples)
         # here, we use the bernoulli probability to sample the anchors
