+++ official/vision/detection/layers/det/rpn.py
@@ -16,11 +16,10 @@ from official.vision.detection.tools.gpu_nms import batched_nms
 
 
 class RPN(M.Module):
-
     def __init__(self, cfg):
         super().__init__()
         self.cfg = cfg
-        self.box_coder = layers.BoxCoder()
+        self.box_coder = layers.BoxCoder(cfg.rpn_reg_mean, cfg.rpn_reg_std)
 
         self.stride_list = cfg.rpn_stride
         rpn_channel = cfg.rpn_channel
@@ -33,12 +32,10 @@ class RPN(M.Module):
         )
         self.rpn_conv = M.Conv2d(256, rpn_channel, kernel_size=3, stride=1, padding=1)
         self.rpn_cls_score = M.Conv2d(
-            rpn_channel, cfg.num_cell_anchors * 2,
-            kernel_size=1, stride=1
+            rpn_channel, cfg.num_cell_anchors * 2, kernel_size=1, stride=1
         )
         self.rpn_bbox_offsets = M.Conv2d(
-            rpn_channel, cfg.num_cell_anchors * 4,
-            kernel_size=1, stride=1
+            rpn_channel, cfg.num_cell_anchors * 4, kernel_size=1, stride=1
         )
 
         for l in [self.rpn_conv, self.rpn_cls_score, self.rpn_bbox_offsets]:
@@ -62,26 +59,32 @@ class RPN(M.Module):
             scores = self.rpn_cls_score(t)
             pred_cls_score_list.append(
                 scores.reshape(
-                    scores.shape[0], 2, self.cfg.num_cell_anchors,
-                    scores.shape[2], scores.shape[3]
+                    scores.shape[0],
+                    2,
+                    self.cfg.num_cell_anchors,
+                    scores.shape[2],
+                    scores.shape[3],
                 )
             )
             bbox_offsets = self.rpn_bbox_offsets(t)
             pred_bbox_offsets_list.append(
                 bbox_offsets.reshape(
-                    bbox_offsets.shape[0], self.cfg.num_cell_anchors, 4,
-                    bbox_offsets.shape[2], bbox_offsets.shape[3]
+                    bbox_offsets.shape[0],
+                    self.cfg.num_cell_anchors,
+                    4,
+                    bbox_offsets.shape[2],
+                    bbox_offsets.shape[3],
                 )
             )
         # sample from the predictions
         rpn_rois = self.find_top_rpn_proposals(
-            pred_bbox_offsets_list, pred_cls_score_list,
-            all_anchors_list, im_info
+            pred_bbox_offsets_list, pred_cls_score_list, all_anchors_list, im_info
         )
 
         if self.training:
             rpn_labels, rpn_bbox_targets = self.get_ground_truth(
-                boxes, im_info, all_anchors_list)
+                boxes, im_info, all_anchors_list
+            )
             pred_cls_score, pred_bbox_offsets = self.merge_rpn_score_box(
                 pred_cls_score_list, pred_bbox_offsets_list
             )
@@ -93,24 +96,26 @@ class RPN(M.Module):
                 rpn_bbox_targets,
                 rpn_labels,
                 self.cfg.rpn_smooth_l1_beta,
-                norm_type="all"
+                norm_type="all",
             )
-            loss_dict = {
-                "loss_rpn_cls": loss_rpn_cls,
-                "loss_rpn_loc": loss_rpn_loc
-            }
+            loss_dict = {"loss_rpn_cls": loss_rpn_cls, "loss_rpn_loc": loss_rpn_loc}
             return rpn_rois, loss_dict
         else:
             return rpn_rois
 
     def find_top_rpn_proposals(
-        self, rpn_bbox_offsets_list, rpn_cls_prob_list,
-        all_anchors_list, im_info
+        self, rpn_bbox_offsets_list, rpn_cls_prob_list, all_anchors_list, im_info
     ):
-        prev_nms_top_n = self.cfg.train_prev_nms_top_n \
-            if self.training else self.cfg.test_prev_nms_top_n
-        post_nms_top_n = self.cfg.train_post_nms_top_n \
-            if self.training else self.cfg.test_post_nms_top_n
+        prev_nms_top_n = (
+            self.cfg.train_prev_nms_top_n
+            if self.training
+            else self.cfg.test_prev_nms_top_n
+        )
+        post_nms_top_n = (
+            self.cfg.train_post_nms_top_n
+            if self.training
+            else self.cfg.test_post_nms_top_n
+        )
 
         batch_per_gpu = self.cfg.batch_per_gpu if self.training else 1
         nms_threshold = self.cfg.rpn_nms_threshold
@@ -125,7 +130,9 @@ class RPN(M.Module):
             batch_level_list = []
             for l in range(list_size):
                 # get proposals and probs
-                offsets = rpn_bbox_offsets_list[l][bid].dimshuffle(2, 3, 0, 1).reshape(-1, 4)
+                offsets = (
+                    rpn_bbox_offsets_list[l][bid].dimshuffle(2, 3, 0, 1).reshape(-1, 4)
+                )
                 all_anchors = all_anchors_list[l]
                 proposals = self.box_coder.decode(all_anchors, offsets)
 
@@ -162,7 +169,9 @@ class RPN(M.Module):
 
             # apply total level nms
             rois = F.concat([proposals, scores.reshape(-1, 1)], axis=1)
-            keep_inds = batched_nms(proposals, scores, level, nms_threshold, post_nms_top_n)
+            keep_inds = batched_nms(
+                proposals, scores, level, nms_threshold, post_nms_top_n
+            )
             rois = rois.ai[keep_inds]
 
             # rois shape (N, 5), info [batch_id, x1, y1, x2, y2]
@@ -181,10 +190,12 @@ class RPN(M.Module):
             batch_rpn_bbox_offsets_list = []
 
             for i in range(len(self.in_features)):
-                rpn_cls_score = rpn_cls_score_list[i][bid] \
-                    .dimshuffle(2, 3, 1, 0).reshape(-1, 2)
-                rpn_bbox_offsets = rpn_bbox_offsets_list[i][bid] \
-                    .dimshuffle(2, 3, 0, 1).reshape(-1, 4)
+                rpn_cls_score = (
+                    rpn_cls_score_list[i][bid].dimshuffle(2, 3, 1, 0).reshape(-1, 2)
+                )
+                rpn_bbox_offsets = (
+                    rpn_bbox_offsets_list[i][bid].dimshuffle(2, 3, 0, 1).reshape(-1, 4)
+                )
 
                 batch_rpn_cls_score_list.append(rpn_cls_score)
                 batch_rpn_bbox_offsets_list.append(rpn_bbox_offsets)
@@ -199,12 +210,10 @@ class RPN(M.Module):
         final_rpn_bbox_offsets = F.concat(final_rpn_bbox_offsets_list, axis=0)
         return final_rpn_cls_score, final_rpn_bbox_offsets
 
-    def per_level_gt(
-        self, gt_boxes, im_info, anchors, allow_low_quality_matches=True
-    ):
+    def per_level_gt(self, gt_boxes, im_info, anchors, allow_low_quality_matches=True):
         ignore_label = self.cfg.ignore_label
         # get the gt boxes
-        valid_gt_boxes = gt_boxes[:im_info[4], :]
+        valid_gt_boxes = gt_boxes[: im_info[4], :]
         # compute the iou matrix
         overlaps = layers.get_iou(anchors, valid_gt_boxes[:, :4])
         # match the dtboxes
@@ -216,7 +225,7 @@ class RPN(M.Module):
         # set negative ones
         labels = labels * (max_overlaps >= self.cfg.rpn_negative_overlap)
         # set positive ones
-        fg_mask = (max_overlaps >= self.cfg.rpn_positive_overlap)
+        fg_mask = max_overlaps >= self.cfg.rpn_positive_overlap
         const_one = mge.tensor(1.0)
         if allow_low_quality_matches:
             # make sure that max iou of gt matched
@@ -224,10 +233,10 @@ class RPN(M.Module):
             num_valid_boxes = valid_gt_boxes.shapeof(0)
             gt_id = F.linspace(0, num_valid_boxes - 1, num_valid_boxes).astype("int32")
             argmax_overlaps = argmax_overlaps.set_ai(gt_id)[gt_argmax_overlaps]
-            max_overlaps = max_overlaps.set_ai(
-                const_one.broadcast(num_valid_boxes)
-            )[gt_argmax_overlaps]
-            fg_mask = (max_overlaps >= self.cfg.rpn_positive_overlap)
+            max_overlaps = max_overlaps.set_ai(const_one.broadcast(num_valid_boxes))[
+                gt_argmax_overlaps
+            ]
+            fg_mask = max_overlaps >= self.cfg.rpn_positive_overlap
         # set positive ones
         _, fg_mask_ind = F.cond_take(fg_mask == 1, fg_mask)
         labels = labels.set_ai(const_one.broadcast(fg_mask_ind.shapeof(0)))[fg_mask_ind]
@@ -258,15 +267,13 @@ class RPN(M.Module):
             num_positive = self.cfg.num_sample_anchors * self.cfg.positive_anchor_ratio
             # sample positive
             concated_batch_labels = self._bernoulli_sample_labels(
-                concated_batch_labels,
-                num_positive, 1, self.cfg.ignore_label
+                concated_batch_labels, num_positive, 1, self.cfg.ignore_label
             )
             # sample negative
             num_positive = (concated_batch_labels == 1).sum()
             num_negative = self.cfg.num_sample_anchors - num_positive
             concated_batch_labels = self._bernoulli_sample_labels(
-                concated_batch_labels,
-                num_negative, 0, self.cfg.ignore_label
+                concated_batch_labels, num_negative, 0, self.cfg.ignore_label
             )
 
             final_labels_list.append(concated_batch_labels)
@@ -279,7 +286,7 @@ class RPN(M.Module):
         self, labels, num_samples, sample_value, ignore_label=-1
     ):
         """ Using the bernoulli sampling method"""
-        sample_label_mask = (labels == sample_value)
+        sample_label_mask = labels == sample_value
         num_mask = sample_label_mask.sum()
         num_final_samples = F.minimum(num_mask, num_samples)
         # here, we use the bernoulli probability to sample the anchors
