+++ official/vision/detection/tools/train.py
@@ -15,7 +15,6 @@ import multiprocessing as mp
 import os
 import sys
 import time
-from collections import defaultdict
 from tabulate import tabulate
 
 import numpy as np
@@ -24,180 +23,17 @@ import megengine as mge
 from megengine import distributed as dist
 from megengine import jit
 from megengine import optimizer as optim
-from megengine.data import Collator, DataLoader, Infinite, RandomSampler
+from megengine.data import DataLoader, Infinite, RandomSampler
 from megengine.data import transform as T
 
 from official.vision.detection.tools.data_mapper import data_mapper
+from official.vision.detection.tools.utils import (
+    AverageMeter,
+    DetectionPadCollator,
+    GroupedRandomSampler
+)
 
 logger = mge.get_logger(__name__)
-
-
-class AverageMeter:
-    """Computes and stores the average and current value"""
-
-    def __init__(self, record_len=1):
-        self.record_len = record_len
-        self.sum = [0 for i in range(self.record_len)]
-        self.cnt = 0
-
-    def reset(self):
-        self.sum = [0 for i in range(self.record_len)]
-        self.cnt = 0
-
-    def update(self, val):
-        self.sum = [s + v for s, v in zip(self.sum, val)]
-        self.cnt += 1
-
-    def average(self):
-        return [s / self.cnt for s in self.sum]
-
-
-def worker(rank, world_size, args):
-    if world_size > 1:
-        dist.init_process_group(
-            master_ip="localhost",
-            master_port=23456,
-            world_size=world_size,
-            rank=rank,
-            dev=rank,
-        )
-        logger.info("Init process group for gpu%d done", rank)
-
-    sys.path.insert(0, os.path.dirname(args.file))
-    current_network = importlib.import_module(os.path.basename(args.file).split(".")[0])
-
-    model = current_network.Net(current_network.Cfg(), batch_size=args.batch_size)
-    params = model.parameters(requires_grad=True)
-    model.train()
-
-    if rank == 0:
-        logger.info(get_config_info(model.cfg))
-    opt = optim.SGD(
-        params,
-        lr=model.cfg.basic_lr * world_size * model.batch_size,
-        momentum=model.cfg.momentum,
-        weight_decay=model.cfg.weight_decay,
-    )
-
-    if args.weight_file is not None:
-        weights = mge.load(args.weight_file)
-        model.backbone.bottom_up.load_state_dict(weights)
-
-    if rank == 0:
-        logger.info("Prepare dataset")
-    loader = build_dataloader(model.batch_size, args.dataset_dir, model.cfg)
-    train_loader = iter(loader["train"])
-
-    for epoch_id in range(model.cfg.max_epoch):
-        for param_group in opt.param_groups:
-            param_group["lr"] = (
-                model.cfg.basic_lr
-                * world_size
-                * model.batch_size
-                * (
-                    model.cfg.lr_decay_rate
-                    ** bisect.bisect_right(model.cfg.lr_decay_stages, epoch_id)
-                )
-            )
-
-        tot_steps = model.cfg.nr_images_epoch // (model.batch_size * world_size)
-        train_one_epoch(
-            model,
-            train_loader,
-            opt,
-            tot_steps,
-            rank,
-            epoch_id,
-            world_size,
-            args.enable_sublinear,
-        )
-        if rank == 0:
-            save_path = "log-of-{}/epoch_{}.pkl".format(
-                os.path.basename(args.file).split(".")[0], epoch_id
-            )
-            mge.save(
-                {"epoch": epoch_id, "state_dict": model.state_dict()}, save_path,
-            )
-            logger.info("dump weights to %s", save_path)
-
-
-def adjust_learning_rate(optimizer, epoch_id, step, model, world_size):
-    base_lr = (
-        model.cfg.basic_lr
-        * world_size
-        * model.batch_size
-        * (
-            model.cfg.lr_decay_rate
-            ** bisect.bisect_right(model.cfg.lr_decay_stages, epoch_id)
-        )
-    )
-    # Warm up
-    if epoch_id == 0 and step < model.cfg.warm_iters:
-        lr_factor = (step + 1.0) / model.cfg.warm_iters
-        for param_group in optimizer.param_groups:
-            param_group["lr"] = base_lr * lr_factor
-
-
-def train_one_epoch(
-    model,
-    data_queue,
-    opt,
-    tot_steps,
-    rank,
-    epoch_id,
-    world_size,
-    enable_sublinear=False,
-):
-    sublinear_cfg = jit.SublinearMemoryConfig() if enable_sublinear else None
-
-    @jit.trace(symbolic=True, opt_level=2, sublinear_memory_config=sublinear_cfg)
-    def propagate():
-        loss_dict = model(model.inputs)
-        opt.backward(loss_dict["total_loss"])
-        losses = list(loss_dict.values())
-        return losses
-
-    meter = AverageMeter(record_len=model.cfg.num_losses)
-    time_meter = AverageMeter(record_len=2)
-    log_interval = model.cfg.log_interval
-    for step in range(tot_steps):
-        adjust_learning_rate(opt, epoch_id, step, model, world_size)
-
-        data_tik = time.time()
-        mini_batch = next(data_queue)
-        data_tok = time.time()
-
-        model.inputs["image"].set_value(mini_batch["data"])
-        model.inputs["gt_boxes"].set_value(mini_batch["gt_boxes"])
-        model.inputs["im_info"].set_value(mini_batch["im_info"])
-
-        tik = time.time()
-        opt.zero_grad()
-        loss_list = propagate()
-        opt.step()
-        tok = time.time()
-
-        time_meter.update([tok - tik, data_tok - data_tik])
-
-        if rank == 0:
-            info_str = "e%d, %d/%d, lr:%f, "
-            loss_str = ", ".join(["{}:%f".format(loss) for loss in model.cfg.losses_keys])
-            time_str = ", train_time:%.3fs, data_time:%.3fs"
-            log_info_str = info_str + loss_str + time_str
-            meter.update([loss.numpy() for loss in loss_list])
-            if step % log_interval == 0:
-                average_loss = meter.average()
-                logger.info(
-                    log_info_str,
-                    epoch_id,
-                    step,
-                    tot_steps,
-                    opt.param_groups[0]["lr"],
-                    *average_loss,
-                    *time_meter.average()
-                )
-                meter.reset()
-                time_meter.reset()
 
 
 def make_parser():
@@ -220,21 +56,6 @@ def make_parser():
     parser.add_argument("--enable_sublinear", action="store_true")
 
     return parser
-
-
-def get_config_info(config):
-    config_table = []
-    for c, v in config.__dict__.items():
-        if not isinstance(v, (int, float, str, list, tuple, dict, np.ndarray)):
-            if hasattr(v, "__name__"):
-                v = v.__name__
-            elif hasattr(v, "__class__"):
-                v = v.__class__
-            elif isinstance(v, functools.partial):
-                v = v.func.__name__
-        config_table.append((str(c), str(v)))
-    config_table = tabulate(config_table)
-    return config_table
 
 
 def main():
@@ -270,6 +91,170 @@ def main():
             p.join()
     else:
         worker(0, 1, args)
+
+
+def worker(rank, world_size, args):
+    if world_size > 1:
+        dist.init_process_group(
+            master_ip="localhost",
+            master_port=23456,
+            world_size=world_size,
+            rank=rank,
+            dev=rank,
+        )
+        logger.info("Init process group for gpu%d done", rank)
+
+    sys.path.insert(0, os.path.dirname(args.file))
+    current_network = importlib.import_module(os.path.basename(args.file).split(".")[0])
+
+    model = current_network.Net(current_network.Cfg(), batch_size=args.batch_size)
+    params = model.parameters(requires_grad=True)
+    model.train()
+
+    if rank == 0:
+        logger.info(get_config_info(model.cfg))
+    opt = optim.SGD(
+        params,
+        lr=model.cfg.basic_lr * world_size * model.batch_size,
+        momentum=model.cfg.momentum,
+        weight_decay=model.cfg.weight_decay,
+    )
+
+    if args.weight_file is not None:
+        weights = mge.load(args.weight_file)
+        model.backbone.bottom_up.load_state_dict(weights)
+
+    if rank == 0:
+        logger.info("Prepare dataset")
+    train_loader = iter(build_dataloader(model.batch_size, args.dataset_dir, model.cfg))
+
+    for epoch_id in range(model.cfg.max_epoch):
+        for param_group in opt.param_groups:
+            param_group["lr"] = (
+                model.cfg.basic_lr
+                * world_size
+                * model.batch_size
+                * (
+                    model.cfg.lr_decay_rate
+                    ** bisect.bisect_right(model.cfg.lr_decay_stages, epoch_id)
+                )
+            )
+
+        tot_steps = model.cfg.nr_images_epoch // (model.batch_size * world_size)
+        train_one_epoch(
+            model,
+            train_loader,
+            opt,
+            tot_steps,
+            rank,
+            epoch_id,
+            world_size,
+            args.enable_sublinear,
+        )
+        if rank == 0:
+            save_path = "log-of-{}/epoch_{}.pkl".format(
+                os.path.basename(args.file).split(".")[0], epoch_id
+            )
+            mge.save(
+                {"epoch": epoch_id, "state_dict": model.state_dict()}, save_path,
+            )
+            logger.info("dump weights to %s", save_path)
+
+
+def train_one_epoch(
+    model,
+    data_queue,
+    opt,
+    tot_steps,
+    rank,
+    epoch_id,
+    world_size,
+    enable_sublinear=False,
+):
+    sublinear_cfg = jit.SublinearMemoryConfig() if enable_sublinear else None
+
+    @jit.trace(symbolic=True, sublinear_memory_config=sublinear_cfg)
+    def propagate():
+        loss_dict = model(model.inputs)
+        opt.backward(loss_dict["total_loss"])
+        losses = list(loss_dict.values())
+        return losses
+
+    meter = AverageMeter(record_len=model.cfg.num_losses)
+    time_meter = AverageMeter(record_len=2)
+    log_interval = model.cfg.log_interval
+    for step in range(tot_steps):
+        adjust_learning_rate(opt, epoch_id, step, model, world_size)
+
+        data_tik = time.time()
+        mini_batch = next(data_queue)
+        data_tok = time.time()
+
+        model.inputs["image"].set_value(mini_batch["data"])
+        model.inputs["gt_boxes"].set_value(mini_batch["gt_boxes"])
+        model.inputs["im_info"].set_value(mini_batch["im_info"])
+
+        tik = time.time()
+        opt.zero_grad()
+        loss_list = propagate()
+        opt.step()
+        tok = time.time()
+
+        time_meter.update([tok - tik, data_tok - data_tik])
+
+        if rank == 0:
+            info_str = "e%d, %d/%d, lr:%f, "
+            loss_str = ", ".join(
+                ["{}:%f".format(loss) for loss in model.cfg.losses_keys]
+            )
+            time_str = ", train_time:%.3fs, data_time:%.3fs"
+            log_info_str = info_str + loss_str + time_str
+            meter.update([loss.numpy() for loss in loss_list])
+            if step % log_interval == 0:
+                average_loss = meter.average()
+                logger.info(
+                    log_info_str,
+                    epoch_id,
+                    step,
+                    tot_steps,
+                    opt.param_groups[0]["lr"],
+                    *average_loss,
+                    *time_meter.average()
+                )
+                meter.reset()
+                time_meter.reset()
+
+
+def get_config_info(config):
+    config_table = []
+    for c, v in config.__dict__.items():
+        if not isinstance(v, (int, float, str, list, tuple, dict, np.ndarray)):
+            if hasattr(v, "__name__"):
+                v = v.__name__
+            elif hasattr(v, "__class__"):
+                v = v.__class__
+            elif isinstance(v, functools.partial):
+                v = v.func.__name__
+        config_table.append((str(c), str(v)))
+    config_table = tabulate(config_table)
+    return config_table
+
+
+def adjust_learning_rate(optimizer, epoch_id, step, model, world_size):
+    base_lr = (
+        model.cfg.basic_lr
+        * world_size
+        * model.batch_size
+        * (
+            model.cfg.lr_decay_rate
+            ** bisect.bisect_right(model.cfg.lr_decay_stages, epoch_id)
+        )
+    )
+    # Warm up
+    if epoch_id == 0 and step < model.cfg.warm_iters:
+        lr_factor = (step + 1.0) / model.cfg.warm_iters
+        for param_group in optimizer.param_groups:
+            param_group["lr"] = base_lr * lr_factor
 
 
 def build_dataset(data_dir, cfg):
@@ -324,99 +309,7 @@ def build_dataloader(batch_size, data_dir, cfg):
         collator=DetectionPadCollator(),
         num_workers=2,
     )
-    return {"train": train_dataloader}
-
-
-class GroupedRandomSampler(RandomSampler):
-    def __init__(
-        self,
-        dataset,
-        batch_size,
-        group_ids,
-        indices=None,
-        world_size=None,
-        rank=None,
-        seed=None,
-    ):
-        super().__init__(dataset, batch_size, False, indices, world_size, rank, seed)
-        self.group_ids = group_ids
-        assert len(group_ids) == len(dataset)
-        groups = np.unique(self.group_ids).tolist()
-
-        # buffer the indices of each group until batch size is reached
-        self.buffer_per_group = {k: [] for k in groups}
-
-    def batch(self):
-        indices = list(self.sample())
-        if self.world_size > 1:
-            indices = self.scatter(indices)
-
-        batch_index = []
-        for ind in indices:
-            group_id = self.group_ids[ind]
-            group_buffer = self.buffer_per_group[group_id]
-            group_buffer.append(ind)
-            if len(group_buffer) == self.batch_size:
-                batch_index.append(group_buffer)
-                self.buffer_per_group[group_id] = []
-
-        return iter(batch_index)
-
-    def __len__(self):
-        raise NotImplementedError("len() of GroupedRandomSampler is not well-defined.")
-
-
-class DetectionPadCollator(Collator):
-    def __init__(self, pad_value: float = 0.0):
-        super().__init__()
-        self.pad_value = pad_value
-
-    def apply(self, inputs):
-        """
-        assume order = ["image", "boxes", "boxes_category", "info"]
-        """
-        batch_data = defaultdict(list)
-
-        for image, boxes, boxes_category, info in inputs:
-            batch_data["data"].append(image)
-            batch_data["gt_boxes"].append(
-                np.concatenate([boxes, boxes_category[:, np.newaxis]], axis=1).astype(
-                    np.float32
-                )
-            )
-
-            _, current_height, current_width = image.shape
-            assert len(boxes) == len(boxes_category)
-            num_instances = len(boxes)
-            info = [
-                current_height,
-                current_width,
-                info[0],
-                info[1],
-                num_instances,
-            ]
-            batch_data["im_info"].append(np.array(info, dtype=np.float32))
-
-        for key, value in batch_data.items():
-            pad_shape = list(max(s) for s in zip(*[x.shape for x in value]))
-            pad_value = [
-                np.pad(
-                    v,
-                    self._get_padding(v.shape, pad_shape),
-                    constant_values=self.pad_value,
-                )
-                for v in value
-            ]
-            batch_data[key] = np.ascontiguousarray(pad_value)
-
-        return batch_data
-
-    def _get_padding(self, original_shape, target_shape):
-        assert len(original_shape) == len(target_shape)
-        shape = []
-        for o, t in zip(original_shape, target_shape):
-            shape.append((0, t - o))
-        return tuple(shape)
+    return train_dataloader
 
 
 if __name__ == "__main__":
